+++ srunner/utilities/code_check_and_formatting.sh
@@ -9,6 +9,6 @@ autopep8 srunner/tools/*.py --in-place --max-line-length=120
 
 pylint --rcfile=.pylintrc --disable=I srunner/scenariomanager/
 pylint --rcfile=.pylintrc srunner/scenarios/
-pylint --rcfile=.pylintrc srunner/challenge/
+#pylint --rcfile=.pylintrc srunner/challenge/
 pylint --rcfile=.pylintrc srunner/tools/
 pylint --rcfile=.pylintrc scenario_runner.py